requirements.txt
# Core LangChain dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.25

# LLM Providers
openai>=1.0.0

# Vector Store and Embeddings
chromadb>=0.4.22
faiss-cpu>=1.7.4  # Alternative: faiss-gpu for GPU support

# Valyu Search (existing tool)
valyu>=0.1.0

# Data handling
pydantic>=2.0.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Logging and monitoring (optional but recommended)
loguru>=0.7.0

# Type hints
typing-extensions>=4.5.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
ruff>=0.1.0
